Convert point-along-line-without-offset test to vitest
The point-along-line-without-offset test was still in the old .cjs
style with top-level throw-based assertions, so vitest reported
"No test suite found". Rewrite it as a describe/it suite using
expect assertions, matching the other converted tests, and remove
the obsolete .cjs file.
